adding ts-node and eslint for ts
in this step the ts-node added i was using nodemon tsc to compile and run the project but it had an error the sloution over the internet and the documents show for dev dep. i have to use ts-node dep. to run the code with nodemon. added eslint plus the file .eslintc.json also added with some custom configaration. in the vscode you have to have a plugin to see the errors in the editor, and in the tsconfig.json we added rootDir and outDir destination.