todoCollection manages todos easily
we have a construtor which includes the username and the given tasks. in the constructor there is a foreach that maps throw todoItems given and set that to a Map called itemMap. For adding a task just with the task name theres a method called addTodo(string) within that function there is a check whcih calls the getTodoById which counts the items and give it a id and set it to the Map with the given task name and id and returns the id. there is a getTodoItems function with boolean includeComplete that returns TodoItems including complete or not. for complete a task and uncomplete there's markComplete method to do that and returns void. remove complete removes the complete items. getItemCount returns total and incomplete numbers